backend/requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
pymongo[srv]
google-genai
python-dotenv
pydantic
